requirements.txt
python==3.11.5
asgiref==3.7.2
Django==5.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==23.5
django-phone-field==1.8.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
pillow==10.3.0
pytest

